chore: add nextest.toml
