+++ packages/taiko-client-rs/crates/proposer/src/proposer.rs
@@ -219,50 +219,3 @@ impl Proposer {
         )))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::borrow::Cow;
-
-    use super::*;
-    use alloy::{
-        primitives::{B256, aliases::U48},
-        rpc::client::NoParams,
-    };
-    use serial_test::serial;
-    use test_context::test_context;
-    use test_harness::ShastaEnv;
-
-    #[test_context(ShastaEnv)]
-    #[serial]
-    #[tokio::test]
-    async fn propose_shasta_batches(env: &mut ShastaEnv) -> anyhow::Result<()> {
-        let proposer = env.proposer.clone();
-        let provider = proposer.rpc_client();
-
-        for i in 0..3 {
-            assert_eq!(B256::ZERO, get_proposal_hash(provider.clone(), U48::from(i + 1)).await?);
-
-            evm_mine(provider.clone()).await?;
-            proposer.fetch_and_propose().await?;
-
-            assert_ne!(B256::ZERO, get_proposal_hash(provider.clone(), U48::from(i + 1)).await?);
-        }
-
-        Ok(())
-    }
-
-    async fn evm_mine(client: ClientWithWallet) -> anyhow::Result<()> {
-        client
-            .l1_provider
-            .raw_request::<_, String>(Cow::Borrowed("evm_mine"), NoParams::default())
-            .await?;
-        Ok(())
-    }
-
-    async fn get_proposal_hash(client: ClientWithWallet, proposal_id: U48) -> anyhow::Result<B256> {
-        let hash: alloy::primitives::FixedBytes<32> =
-            client.shasta.inbox.getProposalHash(proposal_id).call().await?;
-        Ok(hash)
-    }
-}
