+++ packages/taiko-client-rs/crates/test-harness/src/helper.rs
@@ -3,7 +3,7 @@ use std::borrow::Cow;
 use alloy::rpc::client::NoParams;
 use alloy_provider::Provider;
 use anyhow::Context;
-use rpc::client::Client;
+use rpc::client::{Client, ClientWithWallet};
 
 /// Mines a single empty L1 block via the connected execution engine.
 pub async fn mine_l1_block<P>(client: &Client<P>) -> anyhow::Result<()>
@@ -16,6 +16,11 @@ where
         .await
         .context("mining L1 block")?;
     Ok(())
+}
+
+/// Mines a single empty L1 block using a wallet-backed client.
+pub async fn evm_mine(client: &ClientWithWallet) -> anyhow::Result<()> {
+    mine_l1_block(client).await
 }
 
 /// Increases L1 time by the specified number of seconds.
