+++ packages/taiko-client-rs/crates/test-harness/src/shasta/mod.rs
@@ -2,4 +2,4 @@ pub mod env;
 pub mod helpers;
 
 pub use env::ShastaEnv;
-pub use helpers::verify_anchor_block;
+pub use helpers::{get_proposal_hash, verify_anchor_block};
