+++ packages/taiko-client-rs/tests/entrypoint.sh
@@ -74,4 +74,4 @@ export TEST_ACCOUNT_PRIVATE_KEY=0x7c852118294e51e653712a81e05800f419141751be58f6
 export TREASURY=0x1670010000000000000000000000000000010001
 
 cargo nextest -v run \
-    --workspace --exclude bindings --all-features
+    --workspace --exclude bindings --all-features --config-file nextest.toml
